Add files via upload
This commit includes the necessary files to publish my resume as a webpage. It contains an HTML file (index.html) for the resume structure, a CSS file (style.css) for styling, and any additional assets like images. The resume is styled with a dark theme and features contact information, a photo, and section layouts for easy readability.